added the function for plotting the axes, origin etc., so it can be imported rather than defined in the notebook...
...makes things less cluttered for the reader
